Add ability to write video from in memory frames
This commit adds video classes backed by an iterable of 'PIL.Image' frames
which can be used to write video to file.

There is one class for iterables (e.g. list of frames) and one for generators.
Since large video will not fit in memory, this ability allows us to read one
frame from the video file, transform the frame and write it to another video
file. Thus there is only need for one frame at a time in memory.

Signed-off-by: Daniel Falk <Daniel.Falk.1@FixedIT.ai>