Add format for ImageDataSet save
When using the ImageDataSet with S3 storage you need to manually specify
the format in the save_args dict despite the suffix in the file. This is
since the save method on the PIL images are called with an open file.
The PIL documentation states the following about the format argument:

"If a file object was used instead of a filename, this parameter should
always be used."

Signed-off-by: Daniel Falk <daniel.falk.1@fixedit.ai>